requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
datasets==2.14.4
dill==0.3.7
donut-python==1.0.9
dotenv==0.9.9
fastapi==0.95.2
fastapi-soap==0.1.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
h11==0.14.0
huggingface-hub==0.30.1
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
joblib==1.4.2
lightning-utilities==0.14.2
lxml==4.9.4
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.3.1
multiprocess==0.70.15
munch==4.0.0
mysql-connector-python==9.2.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.1.0
platformdirs==4.3.7
propcache==0.3.1
protobuf==6.30.2
pyarrow==19.0.1
pydantic==1.10.21
pydantic-xml==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
sconf==0.2.5
sentencepiece==0.2.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.13.1
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchmetrics==1.7.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.3
triton==3.2.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
xxhash==3.5.0
yarl==1.18.3
zeep==4.3.1
zss==1.2.0
